日志添加删除功能,日志非等高cellbug联调,UI细节调整优化

+++ OpticalValleyUnite/Classes/UI/Journal/ViewController/YQJournalViewController.swift
@@ -246,9 +246,9 @@ extension YQJournalViewController : UITableViewDelegate,UITableViewDataSource{
     
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
         
-        let height = heightDic["\(indexPath.row)"] as! Float
+        let height = heightDic["\(indexPath.row)"] as! CGFloat
         
-        return CGFloat(height)
+        return height
         
     }
